Refactor code into modules
Refactor code by moving functions and tests from `src/main.rs` to new modules and test files.

* **Add `src/parsers.rs`**: Move parsing functions and related tests from `src/main.rs`.
* **Add `src/utils.rs`**: Move enum definitions and utility functions from `src/main.rs`.
* **Add `src/tests/parsers_tests.rs`**: Move parsing function tests from `src/main.rs`.
* **Add `src/tests/utils_tests.rs`**: Move utility function tests from `src/main.rs`.
* **Modify `src/main.rs`**: Remove parsing functions, enum definitions, utility functions, and related tests. Update imports to include new modules `parsers` and `utils`.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/meawoppl/visilog?shareId=XXXX-XXXX-XXXX-XXXX).
